+++ scanpipe/models.py
@@ -562,6 +562,16 @@ class Project(UUIDPKModel, ExtraDataFieldMixin, UpdateMixin, models.Model):
     notes = models.TextField(blank=True)
     settings = models.JSONField(default=dict, blank=True)
     labels = TaggableManager(through=UUIDTaggedItem)
+    purl = models.CharField(
+        max_length=2048,
+        blank=True,
+        help_text=_(
+            "Package URL (PURL) for the project, required for pushing the project's "
+            "scan result to FederatedCode. For example, if the input is an input URL "
+            "like https://registry.npmjs.org/lodash/-/lodash-4.17.21.tgz, the "
+            "corresponding PURL would be pkg:npm/lodash@4.17.21."
+        ),
+    )
 
     objects = ProjectQuerySet.as_manager()
 
@@ -705,6 +715,7 @@ class Project(UUIDPKModel, ExtraDataFieldMixin, UpdateMixin, models.Model):
         """Clone this project using the provided ``clone_name`` as new project name."""
         new_project = Project.objects.create(
             name=clone_name,
+            purl=self.purl,
             settings=self.settings if copy_settings else {},
         )